+++ CHANGELOG.md
@@ -14,6 +14,8 @@ We haven't had the best track record of code/feature changes before this date, b
 - [fb-babel-plugin-utils versions](#fb-babel-plugin-utils)
 
 ### Top-level Github
+- Make demo-app rely on the local versions of our published npm modules.
+  That'll avoid us from having to update `package.json` for every package update.
 - Fix demo-app breakage
 - Require node v10.4.0+ because for fb-tiger-hash (we need BigInt!)
 - *[Experimental]* Add support for Flow type checks.